fix(batch): AI-gate 451 path + Long Undo Snackbar + preview tests (#1048)
* fix(batch): distinguish AI-gate 451 from parse failure + lengthen Undo Snackbar

BatchPreviewViewModel previously surfaced every parse exception as a
generic "Failed to parse batch command" message, so the 451 from
AiFeatureGateInterceptor (master AI toggle off) read like a backend
bug. Adds BatchPreviewErrorKind and routes HttpException(451) to a
"AI features are off — enable them in Settings" copy with no Retry
button (retrying without flipping the toggle just round-trips back
to 451).

Also bumps the post-apply Undo Snackbar from Short (~4s) to Long
(~10s) at both host sites (Today + TaskList) — the 4s default was
un-clickable on the audit's recorded sessions (Phase 1.2 of the
BatchPreview audit). 30s would need a custom snackbar host; deferred.

Adds BatchPreviewViewModelTest covering Ready / EmptyMutations /
AiGate451 / Network / ParseFailure / approve transitions, plus a
Compose smoke test that exercises the screen with mocked repository
state for Ready, Error, and Cancel paths.

* fix(test): use Turbine for approve event assertion in BatchPreviewViewModelTest

The previous shape used backgroundScope.launch + a captured-var collector
to observe BatchEvent.Approved. Under StandardTestDispatcher the
collector subscribed AFTER approve()'s viewModelScope coroutine had
already run emit(), and since _events is MutableSharedFlow(replay = 0)
the event was lost. Turbine's test() block owns the subscription
lifecycle, so awaitItem() pumps the dispatcher with the subscription
already active before approve() is called.

Failure log: PR #1048 lint-and-test
  BatchPreviewViewModelTest > approve_success_emitsApprovedEventAndNotifiesUndoBus
  java.lang.AssertionError: Approve must emit BatchEvent.Approved

---------

+++ app/src/main/java/com/averycorp/prismtask/ui/screens/tasklist/TaskListScreen.kt
@@ -194,10 +194,15 @@ fun TaskListScreen(
             } else {
                 "${event.appliedCount} changes applied"
             }
+            // Long (~10s) is the floor for the batch-undo affordance — the
+            // 4s default Short window made the Undo action effectively
+            // un-clickable on the audit's recorded sessions (Phase 1.2 of
+            // the BatchPreview audit). 30s would need a custom snackbar
+            // host; revisit if Avery requires it.
             val result = viewModel.snackbarHostState.showSnackbar(
                 message = msg,
                 actionLabel = "Undo",
-                duration = SnackbarDuration.Short
+                duration = SnackbarDuration.Long
             )
             if (result == SnackbarResult.ActionPerformed) {
                 batchUndoListener.undo(event.batchId)
